feat: shopId, 날짜로 매출 조회 기능 구현

+++ src/main/java/uos/uos25/employee/entity/Employee.java
@@ -122,7 +122,7 @@ public class Employee {
         if (shop != null) this.shop = shop;
     }
 
-    public Long calculateSalary(Long totalWorkingHour) {
-        return totalWorkingHour * salary;
+    public Integer calculateSalary(Long totalWorkingHour) {
+        return totalWorkingHour.intValue() * salary;
     }
 }
